# swagger-spec.json
{"openapi":"3.0.0","paths":{"/neighborhoods/city/{country}/{state}/{city}":{"get":{"operationId":"NeighborhoodsController_getNeighborhoodsByCity","summary":"","description":"Get Neighborhoods By City Reference","parameters":[{"name":"country","required":true,"in":"path","description":"The Country name of the neighborhood","schema":{"type":"string"}},{"name":"state","required":true,"in":"path","description":"The State name of the neighborhood","schema":{"type":"string"}},{"name":"city","required":true,"in":"path","description":"The City name of the neighborhood","schema":{"type":"string"}}],"responses":{"200":{"description":"Neighborhoods found.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NeighborhoodByCity"}}}}}},"tags":["neighborhoods"]}},"/neighborhoods/state/{country}/{state}":{"get":{"operationId":"NeighborhoodsController_getNeighborhoodsByState","summary":"","description":"Get Neighborhoods By State Reference","parameters":[{"name":"country","required":true,"in":"path","description":"The Country name of the neighborhood","schema":{"type":"string"}},{"name":"state","required":true,"in":"path","description":"The State name of the neighborhood","schema":{"type":"string"}},{"name":"city","required":true,"in":"path","description":"The city of the Neighborhood","schema":{"type":"string"}}],"responses":{"200":{"description":"Neighborhoods found.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NeighborhoodsByState"}}}}},"tags":["neighborhoods"]}},"/cities/state/{country}/{state}":{"get":{"operationId":"CitiesController_getCitiesByState","summary":"","description":"Get Cities By State Reference","parameters":[{"name":"country","required":true,"in":"path","description":"The Country name of the city","schema":{"type":"string"}},{"name":"state","required":true,"in":"path","description":"The State name of the city","schema":{"type":"string"}}],"responses":{"200":{"description":"Cities found.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CityResponse"}}}}}},"tags":["cities"]}},"/cities/country/{country}":{"get":{"operationId":"CitiesController_getCitiesByCountry","summary":"","description":"Get Cities By Country Reference","parameters":[{"name":"country","required":true,"in":"path","description":"The Country name of the city","schema":{"type":"string"}}],"responses":{"200":{"description":"Cities found.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/CitiesByCountry"}}}}},"tags":["cities"]}},"/states/country/{country}":{"get":{"operationId":"StatesController_getStatesByCountry","parameters":[{"name":"country","required":true,"in":"path","description":"The Country name of the state","schema":{"type":"string"}}],"responses":{"200":{"description":"States found.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/StatesByCountry"}}}}}},"tags":["states"]}},"/countries":{"get":{"operationId":"CountriesController_getAllCountries","summary":"","description":"Get All Countries","parameters":[],"responses":{"200":{"description":"Countries found.","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/CountryResponse"}}}}}},"tags":["countries"]}}},"info":{"title":"Places","description":"Places API","version":"0.0.5","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"NeighborhoodByCity":{"type":"object","properties":{"name":{"type":"string","description":"The name of the Neighborhood","example":"Alto Paraná"},"cityId":{"type":"number","description":"The id reference of the City in database","example":1},"city":{"type":"string","description":"The name of the City's Neighborhood","example":"Orleans"},"stateId":{"type":"number","description":"The id reference of the state in database","example":2014},"countryId":{"type":"number","description":"The id reference of the country in database","example":31}},"required":["name","cityId","city","stateId","countryId"]},"NeighborhoodsByState":{"type":"object","properties":{}},"CityResponse":{"type":"object","properties":{"id":{"type":"number","description":"City Id reference in database","example":1},"name":{"type":"string","description":"The name of the City","example":"Orleans"},"countryCode":{"type":"string","description":"The code(ISO3) of the Country's City","example":"BRA"},"countryId":{"type":"number","description":"Country Id reference in database","example":31},"stateCode":{"type":"string","description":"The code(ISO2) of the State's City","example":"SC"},"stateId":{"type":"number","description":"State Id reference in database","example":2014}},"required":["id","name","countryCode","countryId","stateCode","stateId"]},"CitiesByCountry":{"type":"object","properties":{}},"StatesByCountry":{"type":"object","properties":{"id":{"type":"number","description":"State Id reference in database","example":1},"name":{"type":"string","description":"The name of the State","example":"Orleans"},"countryCode":{"type":"string","description":"The code(ISO3) of the Country's State","example":"BRA"},"countryId":{"type":"number","description":"Country Id reference in database","example":31},"stateCode":{"type":"string","description":"The code(ISO2) of the State","example":"SC"}},"required":["id","name","countryCode","countryId","stateCode"]},"CountryResponse":{"type":"object","properties":{"id":{"type":"number","description":"Country Id reference in database","example":1},"name":{"type":"string","description":"Country Name","example":"Brasil"},"iso3":{"type":"string","description":"The code(ISO3) of the Country","example":"BRA"},"iso2":{"type":"string","description":"The code(ISO2) of the Country","example":"BR"},"capital":{"type":"string","description":"Capital","example":"Brasília"},"currency":{"type":"string","description":"Currency Symbol","example":"R$"},"region":{"type":"string","description":"Continent Region of the country","example":"South America"},"subregion":{"type":"string","description":"Continent SubRegion of the country","example":"South"},"alias":{"description":"Alias names","example":["BR","BRA","Brazil","Brasil"],"type":"array","items":{"type":"string"}},"phoneCode":{"type":"number","description":"Phone number code DDI","example":55}},"required":["id","name","iso3","iso2","capital","currency","region","subregion","alias","phoneCode"]}}}}